fix: move factory into separate file
